Automatically render page HtML as it's navigated
Make it more obvious that the browser is actually running: unless --headless is passed in, a new browser tab page will be opened. If a URL was specified, it will be navigated and its HTML will be rendered in the console.

As the tab page is navigated, the HTML will continue to be rendered.