Align dependency resolution with supported nuget/SDK way
Instead of trying to workaround the supported stuff, just align with it. By properly referencing the actual packages (and versions) we want at install/run time, we get all the built-in deps.json generation fixed and have little customization targets to do.

We do fix a targets import timing issue we had in our Directory.Build.props since the Packages.props target was imported without properly defaulting Version, as it would happen for every other target/csproj. This was fixed upstream in devlooped/oss.

Also, we build the runtimes separately so that they exist in the (added) restore sources path by the time we restore/pack the dependant.

When running locally, we just hardcode a published version since it's not really that important.